Moved handleSubmit to Unique Login Function File
